bug fixing - create template hmework
