+++ src/mocks/const.js
@@ -1,0 +1,4 @@
+export const TRANSPORTS = [`taxi`, `bus`, `train`, `ship`, `transport`, `drive`, `flight`];
+export const PLACES = [`check-in`, `restaurant`, `sightseeing`];
+export const TYPES = TRANSPORTS.concat(PLACES);
+export const CITIES = [`London`, `Glasgow`, `Belfast`, `Leeds`, `York`];
