+++ src/view/menu.js
@@ -1,7 +1,7 @@
-import {generateMenuItems} from "../mocks/menu";
+import {generateMenuItems} from "../mocks/menu.js";
 
-const generateMenuMarkup = generateMenuItems().map((el) => {
-  let {title, isChecked} = el;
+const generateMenuMarkup = generateMenuItems().map((element) => {
+  const {title, isChecked} = element;
   return (
     `<a class="trip-tabs__btn ${isChecked ? `trip-tabs__btn--active` : ``}" href="#">${title}</a>`
   );
