+++ src/view/new-point.js
@@ -1,6 +1,3 @@
-
-// add new event
-
 export const createNewPointTemplate = () => {
   return (
     `<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>`
