+++ src/view/sort.js
@@ -1,5 +1,3 @@
-// list-sort.html
-
 export const createSortTemplate = () => {
   return (
     `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
