+++ src/view/trip-info.js
@@ -1,4 +1,3 @@
-
 export const createTripInfoTemplate = () => {
   return (
     `<section class="trip-main__trip-info  trip-info">
